# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for managing tasks.",
        "title": "Task Management API",
        "contact": {
            "name": "Saran",
            "url": "github.com/saran-crayonte/",
            "email": "saran.kumaresan@crayonte.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v2/taskAssignment": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve all task assignments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Assignment"
                ],
                "summary": "Get all task assignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task Assignment retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TaskAssignment"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new task assignment with provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Assignment"
                ],
                "summary": "Create a new task assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Task assignment details",
                        "name": "taskAssignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaskAssignment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task assignment created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Username doesn't exist / Task not found / Task is already assigned",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v2/taskAssignment/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a task assignment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Assignment"
                ],
                "summary": "Get a task assignment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task Assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task assignment retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TaskAssignment"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task assignment not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing task assignment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Assignment"
                ],
                "summary": "Update a task assignment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Updated task assignment details",
                        "name": "taskAssignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaskAssignment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task assignment updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Username doesn't exist / Task not found / Task assignment not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing task assignment by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task Assignment"
                ],
                "summary": "Delete a task assignment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task Assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task assignment deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task assignment not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.TaskAssignment": {
            "type": "object",
            "properties": {
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "startDate": {
                    "type": "string"
                },
                "taskid": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}